vggh

+++ DAL/Request_table.cs
@@ -26,7 +26,7 @@ namespace DAL
         public string Seller_info { get; set; }
         public int User_Id { get; set; }
         public string reqState { get; set; }
-        public int National_ID { get; set; }
+        public long National_ID { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Admin_Req_App_table> Admin_Req_App_table { get; set; }
